refactor: changed upload UI multiple files (#50)
