+++ src/components/StudyRoom/TaskConfirmSheet/TaskConfirmSheet.tsx
@@ -2,45 +2,24 @@ import Image from "next/image";
 
 import { Sheet, SheetContent, SheetHeader } from "@/components/common/Sheet/Sheet";
 
+import { useMultiImageUpload } from "@/hooks/common/useMultiImageUpload";
+
 import type { CommonSheetProps } from "@/types/common";
 import type { TaskConfirmUpdateHandlerType } from "@/types/taskConfirm";
 
 interface TaskConfirmSheetProps extends CommonSheetProps {
-  handleImageUpload: (files: FileList | null) => Promise<string[]>;
   updateInputValue: TaskConfirmUpdateHandlerType;
-  confirmImg: string;
   addItems: string[];
-  itemsType: string[];
-  isAdditem: boolean;
 }
 
 const TaskConfirmSheet = ({
   isOpen,
   onInteractOutside,
-  handleImageUpload,
   updateInputValue,
   addItems,
-  itemsType,
-  isAdditem,
 }: TaskConfirmSheetProps) => {
-  const handleFileChange = async (e: React.ChangeEvent<HTMLInputElement>) => {
-    {
-      /* TODO : 이미지 업로드 공통 hook 사용 */
-    }
-    if (isAdditem) {
-      const images = addItems.concat(await handleImageUpload(e.target.files));
-      const types = itemsType.concat(e.target.id);
-      if (images.length <= 4) {
-        updateInputValue("addItems", images);
-        updateInputValue("itemsType", types);
-      }
-    } else {
-      const uploadImg = (await handleImageUpload(e.target.files))[0];
-      const uploadType = [e.target.id];
-      updateInputValue("confirmImg", uploadImg);
-      updateInputValue("itemsType", uploadType);
-    }
-  };
+  const MAX_SIZE = 4;
+  const { handleImageUpload } = useMultiImageUpload({ maxSize: MAX_SIZE });
 
   return (
     <Sheet open={isOpen}>
@@ -65,11 +44,13 @@ const TaskConfirmSheet = ({
               </label>
               <span>앨범</span>
               <input
-                onChange={handleFileChange}
-                type="file"
-                multiple
-                accept="image/*"
                 id="album"
+                type="file"
+                accept="image/*"
+                multiple
+                onChange={async (e) =>
+                  updateInputValue("addItems", await handleImageUpload(addItems, e.target.files))
+                }
                 hidden
               />
             </div>
