+++ src/hooks/task/useTaskConfirmForm.ts
@@ -9,31 +9,10 @@ interface useTaskConfirmFormProps {
 export const useTaskConfirmForm = ({ initialData }: useTaskConfirmFormProps) => {
   const [taskConfirmForm, setTaskConfirmForm] = useState<TaskConfirmFormType>(
     initialData ?? {
-      confirmImg: "",
       addItems: [],
-      itemsType: [],
       contents: "",
     },
   );
-
-  const getBase64 = (file: File): Promise<string> =>
-    new Promise((resolve, reject) => {
-      const reader = new FileReader();
-
-      reader.readAsDataURL(file);
-
-      reader.onload = () => resolve(reader.result as string);
-
-      reader.onerror = () => reject("");
-    });
-
-  const handleImageUpload = async (files: FileList | null): Promise<string[]> => {
-    if (files) {
-      const images = await Promise.all(Array.from(files).map((image) => getBase64(image)));
-      return images.filter((image) => image !== "");
-    }
-    return [];
-  };
 
   const updateTaskConfirmForm: TaskConfirmUpdateHandlerType = useCallback((key, value) => {
     setTaskConfirmForm((prevTaskForm) => {
@@ -49,5 +28,5 @@ export const useTaskConfirmForm = ({ initialData }: useTaskConfirmFormProps) => 
     console.log(taskConfirmForm);
   };
 
-  return { taskConfirmForm, updateTaskConfirmForm, handleImageUpload, handleSubmit };
+  return { taskConfirmForm, updateTaskConfirmForm, handleSubmit };
 };
